requirements.txt
numpy==1.21.1
Werkzeug==2.0.2
Flask==2.0.2
opencv_python==4.2.0.34
Pillow==8.4.0
tensorflow==2.3.0
gunicorn==19.9.0
scikit-learn==1.0.1
